Add comments from PR

+++ src/IMarketManager.sol
@@ -2,15 +2,41 @@
 pragma solidity ^0.8.13;
 
 interface IMarketManager {
-  function registerMarket(address market) external returns (uint);
-  function setSupplyTarget(uint marketId, uint fundId, uint amount) external;
-  function supplyTarget(uint marketId) external returns (uint);
-  function setLiquidity(uint marketId, uint fundId, uint amount) external returns (uint);
-  function liquidity(uint marketId) external returns (uint);
-  function fundBalance(uint marketId, uint fundId) external returns (int);
-  function totalFundBalance(uint marketId) external returns (int);
-  // What should be the unit here?
-  function deposit(uint marketId, uint amount) external;
-  // amount: amount of synth to redeem.
-  function withdraw(uint marketId, uint amount, address recipient) external;
+    function registerMarket(address market) external returns (uint256);
+
+    // need to think about supply vs liquidity (maybe quantity vs quantity * price)
+    function setSupplyTarget(
+        uint256 marketId,
+        uint256 fundId,
+        uint256 amount
+    ) external;
+
+    function supplyTarget(uint256 marketId) external returns (uint256);
+
+    function setLiquidity(
+        uint256 marketId,
+        uint256 fundId,
+        uint256 amount
+    ) external returns (uint256);
+
+    function liquidity(uint256 marketId) external returns (uint256);
+
+    // fund's debt incurred by underwriting a `buy` for `resBTC`
+    function fundBalance(uint256 marketId, uint256 fundId)
+        external
+        returns (int256);
+
+    // fund's debts across all markets
+    function totalFundBalance(uint256 marketId) external returns (int256);
+
+    // What should be the unit here?
+    // manager takes resUSD from a market, then update debt balances for the funds that setLiquidity to this market and proportionally by the liquidity the fund provided
+    function deposit(uint256 marketId, uint256 amount) external;
+
+    // amount: amount of synth to redeem.
+    function withdraw(
+        uint256 marketId,
+        uint256 amount,
+        address recipient
+    ) external;
 }
